+++ src/diffusers/loaders/controlnet.py
@@ -134,4 +134,5 @@ class FromOriginalControlNetMixin:
         if torch_dtype is not None:
             controlnet = controlnet.to(torch_dtype)
 
+        controlnet.eval()
         return controlnet
