+++ src/diffusers/utils/__init__.py
@@ -84,7 +84,7 @@ from .import_utils import (
     is_xformers_available,
     requires_backends,
 )
-from .loading_utils import load_image
+from .loading_utils import is_single_file_checkpoint, load_image
 from .logging import get_logger
 from .outputs import BaseOutput
 from .peft_utils import (
